AQUILONAQD-955: JENKINS-PIPELINE-MIGRATION
As part of Jenkins Jobs migration to pipeline, changes made to aqd
solution, so build and test can be run as part of train.yaml. Also this
set-up will avoid using gerrit for code review and other purposes which
were followed in the past.

* Script added in train.yaml to get recent TAG from /ms/dev which is
used to generate the documents.
* Upgraded Python version to 2.7.18-64
* Makefile under doc updated so that the recent TAG is not removed.
* /var/tmp used in jenkinstest.sh since the AF_UNIX path is too long if
BASEDIR is used which caused train test to fail on Jenkins.