DDI-1847: Integrate add_interface_address with Infoblox.
Based on Venki's work, the add_interface_address command invokes an add_host REST call to Infoblox.

This commit provides a new concept of "isolated" tests, which do not run as a set of dependencies (the case for existing integration tests, which don't integrate with dsdb and mock it with a file), and use a "known good" database before each test class runs.

Similar to the existing integration tests, the isolated tests launch an AQ broker and an run an Infoblox HTTP mock endpoint in the same Python VM.

In order to make the current integration tests run, an Infoblox HTTP mock endpoint is also started as a daemon process and wired into the dependency chain copying the pattern for launching the AQ broker.

References to the logging framework have been introduced to various test code, and coloured logs used to improve the readability of test output from the isolated tests. They can be run as follows:

$ train test -t isolated