Fix rollback logic of dsdb del_host calls
When the aqd broker sends a `add_host` call to dsdb, the dsdb call can include the `-manager_grn` parameter to dsdb.

If at a future time a `delete_host` call is made to dsdb, the given host is correctly deleted from dsdb. In the event that this `delete_host` call needs to be rolled back, a new `add_host` call is made to dsdb, however, in this rollback case, the `-manager_grn` parameter was not being used, resulting in the host being recreated in dsdb using different parameters than the original host creation.

Addresses-Issue: AQUILONAQD-1726