Fix test restart functionality
The ability to restart tests relies on detecting whether or not a test failed at tearDown time.  To the best of my knowledge, Python's UnitTest class does not provide an interface to detect whether the test failed, hence we have had to rely on side effects to detect this.  With the newer versions of UnitTest, those side effects changed and the way to detect whether or not a test has failed needs to be updated.

Addresses-Issue: AQUILONAQD-2212