Refactor for compatibility with sqlalchemy 1.4
The changes in this PR do not yet upgrade the sqlalchemy dependency from 1.3 to 1.4, however there are a number of changes which address issues when upgrading to 1.4, so code is re-written in a way as to be compatible with both 1.3 and 1.4.  There are further issues which are being addressed, however there is no need to wait for everything to be fixed before pushing working changes out.

Addresses-Issue: AQUILONAQD-2337