Prevent rollback requests for unexecuted ib_services actions
Fixes issue where unexecuted ib_services actions were added to rollback queue. Previously, if an ib_services action (like add_a_record) was not executed (for instance, when the domain was reported as non-authoritative), rollback instructions were still added to the rollback queue. This caused problems if a subsequent request failed (like an add_ptr call), as the system would attempt to rollback the unexecuted add_a_record action. This commit ensures that nothing gets added to the rollback queue if the action was not executed.

Addresses-Issue: AQUILONAQD-2566