Stop running tests on first failure
The aqd broker tests are too interdependent to the point where if one fails, it is virtually certain more will fail, not because of whatever code change caused the first failure, but because of the first failure itself.

Stopping the tests at first failure improves visibility of what the problem that is causing tests to fail is.

Addresses-Issue: AQUILONAQD-2806